moves binarize_extract_bridge to modules
